Fix: model 및 entity 데이터 클래스 속성 타입 변경

+++ app/src/main/java/com/soop_assignment/app/domain/model/SpecificRepo.kt
@@ -3,8 +3,8 @@ package com.soop_assignment.app.domain.model
 data class SpecificRepo(
     val userImageUrl: String,
     val repositoryName: String,
-    val stars: Int,
-    val watchers: Int,
-    val forks: Int,
+    val stars: Long,
+    val watchers: Long,
+    val forks: Long,
     val description: String
 )
